(refactoring): create errs pkg
errs package is responsible to declare errors from grelay
